refactor:move reusable components to seperate directory
